+++ src/main/java/com/global/api/entities/OrderDetails.java
@@ -5,6 +5,7 @@ import lombok.Setter;
 import lombok.experimental.Accessors;
 
 import java.math.BigDecimal;
+import java.util.List;
 
 @Accessors(chain = true)
 @Getter
@@ -15,4 +16,22 @@ public class OrderDetails {
     public boolean hasInsurance;
     public BigDecimal handlingAmount;
     public String description;
+
+    private String localTaxPercentage;
+    private String buyerRecipientName;
+    private String stateTaxIdReference;
+    private String merchantTaxIdReference;
+    private List<OrderDetails.Tax> taxes;
+
+    @Getter
+    @Setter
+    public static class Tax {
+        private String type;
+        private String amount;
+
+        public Tax(String type, String amount) {
+            this.type = type;
+            this.amount = amount;
+        }
+    }
 }