+++ src/main/java/com/global/api/terminals/upa/interfaces/UpaTcpInterface.java
@@ -93,7 +93,6 @@ public class UpaTcpInterface extends DeviceCommInterface {
                                 throw new MessageException("Device is busy.");
                             case Constants.READY_MESSAGE: {
                                 readyReceived = true;
-                                responseMessage = trimControlCodes(response);
                             } break;
                             case Constants.DATA_MESSAGE: {
                                 // Set readyReceived for reboot messages