+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Latest Version - V15.1.2 (02/05/2026)
+### Enhancements:
+[Portico] - Add GatewayTxnId for Interac DebitReversal.
+[GPAPI] - Added all required fields for getting level2 data from GP-API.
+[10376] - Logging Enhancement : provided way to push sdk logs to client logger.
+[UPA] - Add 9F66 and PosSequenceNbr. SequenceNo is now 5F34 only.
+
+### Bug Fixes
+[UPA] - reverted code added for sendReady command failure as this was fixed while updating the UPA version.
+
 ## Latest Version - V15.1.1 (01/15/2026)
 ### Bug Fixes
 - Resolved method mismatch issue for setEnableLogging method from the Configuration class.